# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform V2 API",
...
    "version": "0.5.0"
  },
  "paths": {
...
    "/v1/assets": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "List assets for scan",
        "description": "Paginated inventory for a completed scan. Filter by `resource_type` optionally.",
        "operationId": "list_assets_v1_assets_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Scan UUID",
              "title": "Scan Id"
            },
            "description": "Scan UUID"
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter e.g. aws_iam_user",
              "title": "Resource Type"
            },
            "description": "Filter e.g. aws_iam_user"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetResponse"
                  },
                  "title": "Response List Assets V1 Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/assets/search": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Search assets",
        "description": "Full-text search over resource properties and tags for a scan. Requires `q` and/or `type` and/or `tag`.",
        "operationId": "search_assets_v1_assets_search_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Scan UUID",
              "title": "Scan Id"
            },
            "description": "Scan UUID"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search",
              "title": "Q"
            },
            "description": "Free-text search"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resource type filter",
              "title": "Type"
            },
            "description": "Resource type filter"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tag filter as key or key=value",
              "title": "Tag"
            },
            "description": "Tag filter as key or key=value"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching assets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetResponse"
                  },
                  "title": "Response Search Assets V1 Assets Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/assets/{resource_id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get single asset",
        "description": "One resource by `resource_id` for the given scan.",
        "operationId": "get_asset_v1_assets__resource_id__get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "scan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Scan UUID",
              "title": "Scan Id"
            },
            "description": "Scan UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Asset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          },
          "404": {
            "description": "Asset not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/assets/{resource_id}/relationships": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Asset relationships",
        "description": "IAM graph edges (e.g. role \u2192 policy attachment) for a resource in the scan.",
        "operationId": "get_asset_relationships_v1_assets__resource_id__relationships_get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "scan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Scan UUID",
              "title": "Scan Id"
            },
            "description": "Scan UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Relationship list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelationshipResponse"
                  },
                  "title": "Response Get Asset Relationships V1 Assets  Resource Id  Relationships Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
    "/v1/compliance/frameworks": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "List compliance frameworks",
        "description": "Available frameworks with mapped policies (currently CIS AWS v6).",
        "operationId": "list_frameworks_v1_compliance_frameworks_get",
        "responses": {
          "200": {
            "description": "Framework list",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FrameworkSummary"
                  },
                  "type": "array",
                  "title": "Response List Frameworks V1 Compliance Frameworks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/compliance/frameworks/{framework_id}/scans/{scan_id}": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "CIS control matrix for scan",
        "description": "Full compliance result: score, per-control status, linked findings. Used by platform-ui CIS tab.",
        "operationId": "get_scan_compliance_v1_compliance_frameworks__framework_id__scans__scan_id__get",
        "parameters": [
          {
            "name": "framework_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Framework Id"
            }
          },
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compliance matrix",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanComplianceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Results not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetResponse": {
        "properties": {
          "resource_id": {
            "type": "string",
            "title": "Resource Id",
            "description": "Stable resource identifier (often ARN)"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type",
            "description": "Normalized type, e.g. aws_s3_bucket"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "provider_type": {
            "type": "string",
            "title": "Provider Type"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties",
            "description": "Bronze-normalized resource properties"
          },
          "tags": {
            "additionalProperties": true,
            "type": "object",
            "title": "Tags"
          },
          "collected_at": {
            "type": "string",
            "title": "Collected At"
          },
          "ingested_at": {
            "type": "string",
            "title": "Ingested At"
          }
        },
        "type": "object",
        "required": [
          "resource_id",
          "resource_type",
          "provider",
          "provider_type",
          "account_id",
          "properties",
          "tags",
          "collected_at",
          "ingested_at"
        ],
        "title": "AssetResponse"
      },
...
      "ControlResultResponse": {
        "properties": {
          "control_id": {
            "type": "string",
            "title": "Control Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "pass | fail | not_assessed | manual | error"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "mapped_policy_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mapped Policy Ids"
          },
          "fail_count": {
            "type": "integer",
            "title": "Fail Count"
          },
          "pass_count": {
            "type": "integer",
            "title": "Pass Count"
          },
          "finding_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Finding Ids"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          },
          "evaluated_at": {
            "type": "string",
            "title": "Evaluated At"
          }
        },
        "type": "object",
        "required": [
          "control_id",
          "status",
          "title",
          "mapped_policy_ids",
          "fail_count",
          "pass_count",
          "finding_ids",
          "evidence",
          "evaluated_at"
        ],
        "title": "ControlResultResponse"
      },
...
      "FrameworkSummary": {
        "properties": {
          "framework_id": {
            "type": "string",
            "title": "Framework Id",
            "description": "e.g. cis_aws_v6"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "version_label": {
            "type": "string",
            "title": "Version Label"
          }
        },
        "type": "object",
        "required": [
          "framework_id",
          "title",
          "provider",
          "version_label"
        ],
        "title": "FrameworkSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RelationshipResponse": {
        "properties": {
          "from_resource_id": {
            "type": "string",
            "title": "From Resource Id"
          },
          "to_resource_id": {
            "type": "string",
            "title": "To Resource Id"
          },
          "relationship_type": {
            "type": "string",
            "title": "Relationship Type"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "from_resource_id",
          "to_resource_id",
          "relationship_type",
          "properties",
          "created_at"
        ],
        "title": "RelationshipResponse"
      },
...
      "ScanComplianceResponse": {
        "properties": {
          "framework_id": {
            "type": "string",
            "title": "Framework Id"
          },
          "framework_title": {
            "type": "string",
            "title": "Framework Title"
          },
          "version_label": {
            "type": "string",
            "title": "Version Label"
          },
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Weighted compliance score 0\u2013100"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary",
            "description": "pass/fail/not_assessed counts"
          },
          "evaluated_at": {
            "type": "string",
            "title": "Evaluated At"
          },
          "controls": {
            "items": {
              "$ref": "#/components/schemas/ControlResultResponse"
            },
            "type": "array",
            "title": "Controls"
          }
        },
        "type": "object",
        "required": [
          "framework_id",
          "framework_title",
          "version_label",
          "scan_id",
          "score",
          "summary",
          "evaluated_at",
          "controls"
        ],
        "title": "ScanComplianceResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "SupabaseJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase `access_token` from POST .../auth/v1/token?grant_type=password. Paste token only in Authorize dialog."
      },
      "DevTenantId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID",
        "description": "Tenant UUID. Only when DEV_TENANT_HEADER_AUTH=true."
      },
      "DevRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Role",
        "description": "tenant_admin | viewer | super_admin. Pair with DevTenantId."
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local dev"
    },
    {
      "url": "https://api.example.com",
      "description": "Staging / production"
    }
  ]
}